Workaround for duplicate control binding issue.
Update issue 56
The proper solution to this would be to rationalise the MIPHandler and Control objects so that only one of them wires the control in to the model. However, as a temporary solution we add a check to the top of Model.addControl() which prevents the control being added if it's already in the list.